rag_chatbot/requirements.txt
streamlit
langchain
langchain-groq
groq
faiss-cpu
python-docx
PyPDF2
python-dotenv
pandas
sentence-transformers